added basic routing with react-router-dom. Pages are rendered out of components from src/pages/.
